Move the IBlockDeviceAPI tests back to reduce the size of this branch
